Add files via upload
I entered a business competition at the end of April, and this is my learning code. I tried to use sk-learn packages solving the questions and finding Logic in an uninformed mass of broken data. In this competition, I emploied a variety of data preprocessing strategies, and used a number of machine learning models.